Update main.js
Add a function called handleResponse that will be called when JSON data is received.
In the form submission event, create a dynamic script and set its src to the API URL with a callback parameter pointing to the handleResponse function.
When the script is loaded, the handleResponse function is invoked with the received JSON data.
In the handleResponse function, create an HTML table and populate it with the JSON data, displaying it on the page.
This technique allows you to work around the CORS error by using JSONP, which is a way to make cross-domain requests. Ensure that the third-party API supports JSONP for this to work.